TVPvariance
Performs LASSO regression in a moving window by using BIC criterion to choose penalty 
parameter (lambda). The simulated data contains a break point after which the 
variance of the error term changes. Plots time series of lambda in LASSO regression. 
Furthermore, the cardinality of the active set q, the L2-norm of the residuals, the 
L1-norm of the paramter beta  and the condition number of the squared design matrix 
X'X) are plotted.